refactor: Rename `lib/plugins/package/lib/zipService.js`
