+++ lib/plugins/package/lib/zip-service.js
