+++ lib/plugins/package/package.js
@@ -4,7 +4,7 @@ const BbPromise = require('bluebird');
 const path = require('path');
 const ServerlessError = require('../../serverless-error');
 const cliCommandsSchema = require('../../cli/commands-schema');
-const zipService = require('./lib/zipService');
+const zipService = require('./lib/zip-service');
 const packageService = require('./lib/package-service');
 
 class Package {
