+++ test/unit/lib/plugins/package/lib/zip-service.test.js
